requirements.txt
pykeepass
ansible-core
